redesigning buffers

+++ LinaCore/include/Platform/Win32/Win32PlatformProcess.hpp
@@ -44,8 +44,8 @@ namespace Lina
 	class Win32PlatformProcess
 	{
 	public:
-		static void SleepSpinLock(double milliseconds);
-		static void Sleep(double milliseconds);
+		static void SleepSpinLock(uint64 milliseconds);
+		static void Sleep(uint64 milliseconds);
 
 	private:
 		friend class Application;
